// src/app/syllabus-entry/models/TrainingSyllabus.ts
export interface TrainingSyllabus {
      trainingSyllabusId:number;
      trainingObjectiveId:number;
      courseTaskId: number;
      baseSchoolNameId: number;
      courseNameId: number;
      courseName: string,
      courseDurationId: number;
      bnaSubjectNameId: number;
      syllabusDetail: string;
      t:number;
      p:number;
      m:number;
      remarks:number;
      menuPosition:number;
      isActive: boolean
}